fix(persistence): resolve sql injection and implement parameter bounds
- Replace LIMIT string interpolation with parameterized placeholders across all Turso storage modules (episodes, patterns, tag operations, monitoring).
- Implement MAX_TAGS_PER_QUERY bound (100) and enforce retrieval limits in tag_operations.rs and monitoring.rs.
- Refactor mod.rs to move internal embedding implementations to embeddings_internal.rs, maintaining <500 LOC per file.
- Add comprehensive unit tests for parameterized queries, migration logic, and input bounds.
- Update progress/LEARNINGS.md with security findings and prevention strategies.

Security: Prevents potential SQL injection via LIMIT clause and resource exhaustion from unbounded tag queries.
